requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==3.2.2
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.124.4
ffmpeg==1.4
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
qrcode==8.2
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.25
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.27.0
watchfiles==1.1.1
websockets==15.0.1
wifi-qrcode-generator==1.0.1
